Refactor the HtmlTextConverter
There were multiple issues with this class that made it hard to
extend, adapt and test.

Not only did the converter take care of converting text, it also
took care of loading and iterating the plain text content.
This made instantiating the converter very inflexible. In order
to Test it, one had to make sure there was a plain text file available
during the test.

This also meant that the only thing that could be converted was a
plain text file. That might not be what's desired in the future.

Several steps were taken to improve the code. Firstly, providing
the converter with Strings to convert has been extracted into
an interface. It seems to make sense for the converter to receive
individual lines to convert. Consequently, the interface provides
it with a Stream of lines.

The file based approach has been extracted into an implementation
of that interface. The converter's test provides a simple mock
implementation.

The new implementation respects several design principles that
weren't respected previously.
* The Single Responsibility Principle was violated by the
  Converter for the reasons described above.
* The converter also ignored the Open-Closed-Principle, since
  it made it virtually impossible to extend the application
  by providing anything but a text file to be converted.
  Now, it can create HTML out of anything capable of supplying
  it with a Stream of lines.
* The Dependency Inversion Principle was also somewhat violated.
  Although the converter didn't actually depend on the FileReader
  or BufferedReader by expecting them in its constructor, it
  relied upon them, nevertheless. By taking care of loading the
  file elsewhere, and constructing the converter with an
  interface, that implicit dependency has been removed.